JDP230301-3 code refactor
